Analyzing American Baby Name Trends
This SQL project is a guided project from DataCamp analyzing a dataset that contains name, sex assigned with each name, year, and the number of times a baby was named that specific name during a given year. My responses to the prompts showcase the most popular male and female baby names over the years through summing, counting, ranking, and use of subqueries and a Common Table Expression.